Add merging multiplexer (#45)
